+++ CHANGELOG.md
@@ -7,12 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Fix missing `#[cfg_attr(...]` in `psp_sys::kernel`
+
 ## [v0.2.1] - 2018-09-12
 ### Added
 - `dox` feature to disable linking to `vitasdk` stubs during compilation (use in  `docs.rs`)
 
 ### Fixed
-
 - Links in `CHANGELOG.md` file.